refactor: move malicious chunk tests to own file
